o patched lockable node/relationship in HA to use long o implemented proper hashCode for objects with long id (Node,Relationship,Commands)

+++ advanced/ha/src/main/java/org/neo4j/kernel/ha/MasterImpl.java
@@ -246,7 +246,7 @@ public class MasterImpl implements Master
         Node[] result = new Node[ids.length];
         for ( int i = 0; i < ids.length; i++ )
         {
-            result[i] = new LockableNode( (int) ids[i] );
+            result[i] = new LockableNode( ids[i] );
         }
         return result;
     }
@@ -256,7 +256,7 @@ public class MasterImpl implements Master
         Relationship[] result = new Relationship[ids.length];
         for ( int i = 0; i < ids.length; i++ )
         {
-            result[i] = new LockableRelationship( (int) ids[i] );
+            result[i] = new LockableRelationship( ids[i] );
         }
         return result;
     }
